+++ internal/backends/postgresql/collection.go
@@ -64,7 +64,7 @@ func (c *collection) Query(ctx context.Context, params *backends.QueryParams) (*
 
 	if p == nil {
 		return &backends.QueryResult{
-			Iter: newQueryIterator(ctx, nil),
+			Iter: newQueryIterator(ctx, nil, nil),
 		}, nil
 	}
 
@@ -75,13 +75,11 @@ func (c *collection) Query(ctx context.Context, params *backends.QueryParams) (*
 
 	if meta == nil {
 		return &backends.QueryResult{
-			Iter: newQueryIterator(ctx, nil),
+			Iter: newQueryIterator(ctx, nil, nil),
 		}, nil
 	}
 
-	d := newDatabase(c.r, c.dbName)
-
-	list, err := d.ListCollections(ctx, new(backends.ListCollectionsParams))
+	list, err := newDatabase(c.r, c.dbName).ListCollections(ctx, new(backends.ListCollectionsParams))
 	if err != nil {
 		return nil, lazyerrors.Error(err)
 	}
@@ -118,20 +116,17 @@ func (c *collection) Query(ctx context.Context, params *backends.QueryParams) (*
 		return nil, lazyerrors.Error(err)
 	}
 
+	var s scanner = new(documentScanner)
 	if params.OnlyRecordIDs {
-		return &backends.QueryResult{
-			Iter: newQueryIteratorWithScanner(ctx, rows, new(onlyRecordIDScanner)),
-		}, nil
+		s = new(onlyRecordIDScanner)
 	}
 
 	if cInfo.Capped() {
-		return &backends.QueryResult{
-			Iter: newQueryIteratorWithScanner(ctx, rows, new(recordIDScanner)),
-		}, nil
+		s = new(recordIDScanner)
 	}
 
 	return &backends.QueryResult{
-		Iter: newQueryIterator(ctx, rows),
+		Iter: newQueryIterator(ctx, rows, s),
 	}, nil
 }
 
@@ -151,9 +146,7 @@ func (c *collection) InsertAll(ctx context.Context, params *backends.InsertAllPa
 		return nil, lazyerrors.Error(err)
 	}
 
-	d := newDatabase(c.r, c.dbName)
-
-	list, err := d.ListCollections(ctx, new(backends.ListCollectionsParams))
+	list, err := newDatabase(c.r, c.dbName).ListCollections(ctx, new(backends.ListCollectionsParams))
 	if err != nil {
 		return nil, lazyerrors.Error(err)
 	}
@@ -173,26 +166,6 @@ func (c *collection) InsertAll(ctx context.Context, params *backends.InsertAllPa
 				return lazyerrors.Error(err)
 			}
 
-			if cInfo.Capped() {
-				q := fmt.Sprintf(
-					`INSERT INTO %s (%s,%s) VALUES ($1,$2)`,
-					pgx.Identifier{c.dbName, meta.TableName}.Sanitize(),
-					metadata.RecordIDColumn,
-					metadata.DefaultColumn,
-				)
-
-				if _, err = tx.Exec(ctx, q, doc.RecordID(), string(b)); err != nil {
-					var pgErr *pgconn.PgError
-					if errors.As(err, &pgErr) && pgErr.Code == pgerrcode.UniqueViolation {
-						return backends.NewError(backends.ErrorCodeInsertDuplicateID, err)
-					}
-
-					return lazyerrors.Error(err)
-				}
-
-				continue
-			}
-
 			// use batches: INSERT INTO %s %s VALUES (?), (?), (?), ... up to, say, 100 documents
 			// TODO https://github.com/FerretDB/FerretDB/issues/3271
 			q := fmt.Sprintf(
@@ -201,7 +174,19 @@ func (c *collection) InsertAll(ctx context.Context, params *backends.InsertAllPa
 				metadata.DefaultColumn,
 			)
 
-			if _, err = tx.Exec(ctx, q, string(b)); err != nil {
+			var args []any
+			if cInfo.Capped() {
+				q = fmt.Sprintf(
+					`INSERT INTO %s (%s,%s) VALUES ($1,$2)`,
+					pgx.Identifier{c.dbName, meta.TableName}.Sanitize(),
+					metadata.RecordIDColumn,
+					metadata.DefaultColumn,
+				)
+				args = append(args, doc.RecordID())
+			}
+
+			args = append(args, string(b))
+			if _, err = tx.Exec(ctx, q, args...); err != nil {
 				var pgErr *pgconn.PgError
 				if errors.As(err, &pgErr) && pgErr.Code == pgerrcode.UniqueViolation {
 					return backends.NewError(backends.ErrorCodeInsertDuplicateID, err)
@@ -352,9 +337,7 @@ func (c *collection) Explain(ctx context.Context, params *backends.ExplainParams
 		}, nil
 	}
 
-	d := newDatabase(c.r, c.dbName)
-
-	list, err := d.ListCollections(ctx, new(backends.ListCollectionsParams))
+	list, err := newDatabase(c.r, c.dbName).ListCollections(ctx, new(backends.ListCollectionsParams))
 	if err != nil {
 		return nil, lazyerrors.Error(err)
 	}
